I have corrected the player life's variable type from byte to int, as it caused errors at certain points in the game when the value turned negative

+++ Assets/Scripts/Player.cs
@@ -11,7 +11,7 @@ public class Player : MonoBehaviour
     //Private attributes
     private Animator _animatorPlayer;
     private SpriteRenderer _spriteRendererPlayer;
-    private byte _livesPlayer;
+    private int _livesPlayer;
 
     //Methods
     private void Start()
